Add entities and move pokemon and session
